Move physics-related functions from engines.py to physics.py
